requirements.txt
redis
pymongo
cassandra-driver
neo4j
requests